add components CartTotals, CartTotalsCheckout
